+++ src/main/java/com/github/gtexpert/gtwp/integration/rthings/recipes/RThingsWoodRecipe.java
@@ -1,0 +1,36 @@
+package com.github.gtexpert.gtwp.integration.rthings.recipes;
+
+import java.util.Arrays;
+import java.util.List;
+
+import gregtech.loaders.WoodTypeEntry;
+
+import com.github.gtexpert.gtwp.api.util.Mods;
+import com.github.gtexpert.gtwp.loaders.GTWPWoodRecipeLoader;
+
+public class RThingsWoodRecipe {
+
+    private static List<WoodTypeEntry> DEFAULT_ENTRIES;
+
+    private static List<WoodTypeEntry> getDefaultEntries() {
+        if (DEFAULT_ENTRIES == null) {
+            final String mcModId = Mods.Names.RANDOM_THINGS;
+            return DEFAULT_ENTRIES = Arrays.asList(
+                    new WoodTypeEntry.Builder(mcModId, "spectre")
+                            .log(Mods.RandomThings.getItem("spectrelog", 1)).removeCharcoalRecipe()
+                            .planks(Mods.RandomThings.getItem("spectreplank", 1), "spectreplank")
+                            .build());
+        }
+        return DEFAULT_ENTRIES;
+    }
+
+    public static void init() {
+        for (WoodTypeEntry entry : getDefaultEntries()) {
+            GTWPWoodRecipeLoader.removePlankRecipe(false, entry, Mods.Names.RANDOM_THINGS);
+
+            GTWPWoodRecipeLoader.registerWoodTypeRecipe(false, entry);
+            GTWPWoodRecipeLoader.addCuttingRecipe(entry);
+            GTWPWoodRecipeLoader.addSawmillRecipe(entry);
+        }
+    }
+}